requirements.txt
django==5.0.2
django-allauth==0.61.1
PyGithub==2.2.0
langchain_openai==0.0.7
langchain==0.1.9
langchain-community==0.0.24
langchainhub==0.1.14
openai==1.10.0
GitPython==3.1.42
instructor==0.3.5
beautifulsoup4==4.12.3
html2text
kubernetes==29.0.0
jinja2==3.1.3
psycopg2-binary==2.9.9
django-tables2==2.7.0
Markdown==3.5.2
sentry-sdk==1.41.0
stripe==8.7.0